+++ samples/json/json_matcher.cpp
@@ -1,5 +1,5 @@
 // lug - Embedded DSL for PE grammar parser combinators in C++
-// Copyright (c) 2017-2024 Jesse W. Towner
+// Copyright (c) 2017-2025 Jesse W. Towner
 // See LICENSE.md file for license details
 
 #include <lug/lug.hpp>
@@ -19,12 +19,12 @@ public:
 		rule Boolean        = lexeme[ "true"_sx | "false" ];
 		rule Null           = lexeme[ "null" ];
 		rule UnicodeEscape  = lexeme[ 'u' > "[0-9A-Fa-f][0-9A-Fa-f][0-9A-Fa-f][0-9A-Fa-f]"_rx ];
-		rule Escape         = lexeme[ "\\" > ("[/\\bfnrt]"_rx | UnicodeEscape) ];
-		rule String         = lexeme[ "\"" > *("[^\"\\\u0000-\u001F]"_rx | Escape) > "\"" ];
-		rule Array          = "[" > JSON > *("," > JSON) > "]";
-		rule Object         = "{" > String > ":" > JSON > *("," > String > ":" > JSON) > "}";
-		JSON                = Object | Array | String | Number | Boolean | Null;
-		grammar_ = start(JSON);
+		rule Escape         = lexeme[ '\\' > ("[/\\bfnrt]"_rx | UnicodeEscape) ];
+		rule String         = lexeme[ '"' > *("[^\"\\\u0000-\u001F]"_rx | Escape) > '"' ];
+		rule Array          = '[' > JSON > *(',' > JSON) > ']';
+		rule Object         = '{' > String > ':' > JSON > *(',' > String > ':' > JSON) > '}';
+		JSON                = Null | Boolean | Number | String | Array | Object;
+		grammar_            = start(JSON > eoi);
 	}
 
 	bool parse_cin()